requirements.txt
Flask
python-dotenv
numpy
librosa
soundfile
matplotlib
tensorflow
